dependencies: Fix parsing of shebangs with spaces
While finding an external program, we should only split the shebang
once since that is what Linux and BSD also do. This is also why
everyone uses #!/usr/bin/env in their shebangs since that allows
you to run an interpreter in a path with spaces in it.

See `man execve` for more details, specifically the sections for
interpreter scripts.