dependencies: Use prefix variable with define_variable
The test used by the new define_variable parameter was using the
includedir directory. However, in order to get a successful test,
the includedir variables needs to be relative to the prefix
variable, otherwise the replacemente will not have any effect.

This changes uses the prefix variable itself because we can
assure that it will be present.