Add 'static_paths' global option
It is a list of path prefixes where it is allowed to use static
libraries. By default it has "/" which means all paths are allowed.
Sometimes it is useful to have there the path to an SDK that we want to
static link and keep dynamic link for everything else (e.g. plateform
libraries).

Closes #2765.