build: Add optional dependencies for RunScripts
Sometimes RunScript are generated as complementary objects to
RunTargets. However, there is no direct relationship between them.

A new dependency attribute has been added to RunScripts, so it
could be known what it the related target. This way, when
executing the RunScript the building of the RunTarget can be
triggered.