Use GObject-based libraries in test case/vala/11 generated vapi
Non-GObject-based libraries are not (reliably) introspectable, and this only
happens to work by accident, at the moment.

Briefly: In non-libtool mode, g-ir-scanner inspects the run-time
dependencies of a generated executable to determine the shared library name
an -l flag corresponds to.  But this generated executable only references
the libraries by *_get_type() functions.  So, if the library doesn't contain
any Gobject-based types, it is not referenced, and this inspection fails
with PECOFF binaries.  It seems that distutils doesn't currently decide to
use --as-needed, so this just happens to work on ELF.