Allow override config-tools from the command line
For dependencies that are found via a config tool there is currently no
good way to force a specific tool to be used. You can set $PATH, but
that can have side effects, or allow falling back to a different tool.
In the case of LLVM where we walk through the list of tools and always
pick the highest version that fulfills any version requirements this is
even harder.

The most obvious solution to me is to add a command line option to force
the use of one specific binary over another, which this patch does. But,
it only applies to native dependencies, for cross compilation the
binaries section of the cross file should be used instead. The reason I
prefer a command line option in this case is that it solves the problem
for all meson based projects and it doesn't require projects to
implement special logic to pass an argument into the dependency call,
like some other options would.

The one caveat to this patch is that I can't figure out a way for this
command line argument to invalidate the dependency cache, which means
that even if someone passes -Dc_llvm_config=/something/different the
dependency wont change, nor will an error be raised. That seems bad to
me, but I don't see any other cache invalidation method for pkg-config,
so I assume that means this patch doesn't change the status quo.

Fixes #2887