[skip ci] Adding a "Projects using Meson" to Community section
Currently the website doesn't seem to contain any info on projects using Meson.

A build system lives or dies by the people who are willing to use the project so I've added a small "Projects using Meson" section to the "Community" section of the website.

I'm sure this list could be greatly improved, you should be promoting your success stories!