[skip ci] Improve Contributing file slightly
[GitHub links to your Contributing instructions when people send pull
requests](https://help.github.com/articles/setting-guidelines-for-repository-contributors/).
For this reason it is good for this file to be a bit nicer and more
welcoming.

This file has been changed to Markdown so GitHub renders it nicely and
some encouragement text has been added. The link to the full
contributing instructions has also been reformatted and this makes the
file look nicer when viewing in plain text form too!