Use sane permission defaults in install_data if not specified
Instead of just copying the permissions from the source file, which can
be affected by umask, just use one of 0644 or 0755 depending on whether
the original file was executable or not.

Source control systems such as git will preserve the executable bit, but
they may make files group-writable or world-readable depending on the
umask settings. That should not affect the permissions of the installed
files though, since they can legitimately vary based on the development
environment.

Tested: with a script that checks out a git repository with different
umasks, run a meson.build with a couple install_data()s and checks that
the resulting installed files were not affected by the umask.

Fixes #3160