gtkdoc: ignore SameFileError
When using the new support for generated content files it can happen
that the generated file is already in the right place (because due to
past hacks going back to automake based gtkdoc that was the way the
project made it all work). Instead of failing we can just ignore that
error.

v2: Fix overindent