Use dependencies for cc.find_library()
Allows the `dependencies' keyword argument to be used with the
cc.find_library() command. This has the effect of adding the compile and
link arguments of each dependency to the command-line which may be
necessary to find the library. The `dirs' keyword argument can still be
used to specify a list of search directories that overrides any of the
dependencies specified.

Note that specifying the list of search directories is not safe unless
you know exactly what you are doing because it will not attempt to link
to the library. Instead it will return the path to the library if it
exists, but that does not automatically guarantee that the library can
really be linked against (for example during cross-compilation). This
can lead the configuration step to succeed but give an error during
linking. Attempting to link against the library using the library paths
provided by the dependencies should therefore be the preferred method.

Signed-off-by: Thierry Reding <treding@nvidia.com>