pkgconfig generator: Add support for Requires.internal
This introduce a few changes:

- Do not differentiate StaticLibrary and SharedLibrary anymore, it was
made only has a temporary hack around the lack of Requires.internal
support in pkg-config.

- CFLAGS from a Dependency passed to libraries_private are now added
into the pc file to respect the same semantic than pkg-config's
Requires.private.

- Dependencies pulled from Library objects are added as internal by
default. It means that if they are Dependency objects their CFLAGS won't
be added. If they are PkgConfigDependency objectsthey are still added to
Requires.private until we can depend on a version of pkg-config that
supports Requires.internal.