compilers: Cache compiler checks
This caching is only for a single run, so it doesn't help reconfigure.
However, it is useful for subproject setups where different subprojects
will run the same compiler checks.

The cache is also per compiler instance.

For gst-build, this halves the number of compiler checks that are run
and reduces configuration time by 20%.