Always generate a new Product GUID. Closes #2485. [skip ci]
We are taking some shortcuts here. The WiX documentation says that you
should keep the Product GUID the same for "small and minor" upgrades
but change it for major ones. These are not defined in any way and a
change of version number might, or might not, warrant a guid
update. For simplicity we will always regenerate the Product GUID.

Again we find that naming things is difficult since "product" in
everyday language would mean "the application/library/software" and
all different versions of it. In MSI installer terminology it means
something vague between the two.

https://www.firegiant.com/wix/tutorial/upgrades-and-modularization/