mesonmain: Add --wipe argument
It is similar to --reconfigure but completely wipe the build directory
first. Command line options are stored in a separate file to be able to
reconfigure with the same options. It is intended to make easier to
rebuild project with a newer version of meson, or when builddir somehow
got corrupted. This means that we should try as much as possible to
maintain backward compatibility for the cmd_line.txt file format, which
is the reason why we cannot use coredata.dat for that.

Fixes #3542.