ninjabackend: generate separate SHSYM rules for each compiler
Instead of passing a $CROSS variable in the "build" directive, define
separate rules.  This introduces some duplication in the rules, but
removes it from the targets because the --cross-host=...  option depends
only on whether the target is native or cross-compiled.  Previously,
it was specified via $CROSS once per shared library or module.

While at it, extract the code function to generate SHSYM rules into a
separate function, and generalize the function that generates SHSYM
"build" directives.  We will use them shortly to generate a list
of undefined symbols in the static library.  That usage in turn will
actually require to have separate rules for each compiler.