FeatureNew: fix and cleanup minimum version comparison.
This could be probably be further cleaned up, as the code
was copy pasted from version_compare, but that would be more
intrusive.

For now simply fix the used comparison operator, along with a
test showing what is fixed (test_using_recent_feature), and
also remove dead code paths in the comparison function.