Add error check for cross-file absolute paths
When an invalid cross-file filename is given we get different error
messages depending on whether it is given by absolute path or not.
Fix this by checking for the presence of absolute path files in the
same function where we check the relative path ones. Also, put the
error message into a variable to avoid multiple copies of the same
constant string.