allow reading cross-file from stdin
Read the cross-file data from standard-in rather than from a file on disk,
and write the read data to the file "meson.crossfile" in the build
directory. This allows projects to programatically generate the cross
file, or more likely, to have standard sets of cross files than can then
be tweaked e.g. using sed, to work in the specific case. For example, it
can save having hard-coded paths to binaries in the cross-file, and those
can be filled in later using sed replacements.